PSARC 2005/373 Contract for the Miniroot Repository 6263254 console output gets graphically corrupted after the newboot 6270368 grub does not honor vendor option 150 to specify menu.lst file 6271324 boot-archive.xml is delivered into the wrong directory 6271395 grub should show progress when loading the root archive from slow devices 6272654 gcc and krtld don't get along 6273069 net install fails on v40z with external 3310 disk array 6280557 serial console redirection does not work right in some corner cases 6280646 amd64 memlist structure mismatch causes panic 6284029 boot may generate supurious warning on disabled pci config space 6285006 pxegrub truncates DHCPREQUEST packet

+++ usr/src/uts/common/krtld/bootrd.c
@@ -65,8 +65,8 @@ diskread(fileid_t *filep)
 	/* add in offset of root slice */
 	blocknum = filep->fi_blocknum;
 
-	diskloc = (caddr_t)rd_start + blocknum * DEV_BSIZE;
-	if (diskloc + filep->fi_count > (caddr_t)rd_end) {
+	diskloc = (caddr_t)(uintptr_t)rd_start + blocknum * DEV_BSIZE;
+	if (diskloc + filep->fi_count > (caddr_t)(uintptr_t)rd_end) {
 		_kobj_printf(ops, "diskread: start = 0x%p, size = 0x%x\n",
 		    diskloc, filep->fi_count);
 		_kobj_printf(ops, "reading beyond end of ramdisk\n");
@@ -132,8 +132,8 @@ bkmem_alloc(size_t size)
 
 	addr = BOP_ALLOC(ops, 0, size, 0);
 #ifdef	DEBUG
-	if (scratch_max < (uint64_t)addr + size)
-		scratch_max = (uint64_t)addr + size;
+	if (scratch_max < (uintptr_t)addr + size)
+		scratch_max = (uintptr_t)addr + size;
 #endif
 	return (addr);
 }
